pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel>=0.45.1"]
build-backend = "setuptools.build_meta"

[project]
name = "comproscanner"
version = "0.1.0"
description = "Multi-agent system for extracting and processing structured composition-property data from scientific literature"
readme = "README.md"
authors = [{ name = "Aritra Roy", email = "contact@aritraroy.live" }]
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "tqdm>=4.67.1",
    "lxml>=5.3.1",
    "pandas>=2.2.3",
    "torch>=2.6.0",
    "langchain>=0.3.19",
    "transformers>=4.49.0",
    "mysql>=0.0.3",
    "mysql-connector>=2.2.9",
    "langchain_community>=0.3.18",
    "chromadb>=0.6.3",
    "pdf2doi @ git+https://github.com/MicheleCotrufo/pdf2doi.git", # pdfminer.six>=20231228 is needed for CrewAI, it is updated in the repo but not released yet by pdf2doi as a package
    "docling>=2.28.0",
    "neo4j>=5.28.1",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Materials Science",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
[project.optional-dependencies]
dev = ["pytest>=8.3.4", "pytest-mock>=3.14.0", "pytest-cov>=6.0.0"]
cuda = [
    "torchaudio>=2.6.0+cu121; platform_system=='Linux' or platform_system=='Windows'",
]
[project.urls]
"Homepage" = "https://github.com/aritraroy24/comproscanner"
"Bug Tracker" = "https://github.com/aritraroy24/comproscanner/issues"
[tool.setuptools]
package-dir = { "" = "src" }
packages = ["comproscanner"]
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
